+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/type/Candidate.kt
@@ -299,14 +299,13 @@ public class FinishReason private constructor(public val name: String, public va
  *
  * If using Grounding with Google Search, you are required to comply with the "Grounding with Google
  * Search" usage requirements for your chosen API provider:
- * [Gemini Developer
- * API](https://ai.google.dev/gemini-api/terms#grounding-with-google-search) or
+ * [Gemini Developer API](https://ai.google.dev/gemini-api/terms#grounding-with-google-search) or
  * Vertex AI Gemini API (see [Service Terms](https://cloud.google.com/terms/service-terms) section
  * within the Service Specific Terms).
  *
  * @property webSearchQueries The list of web search queries that the model performed to gather the
  * grounding information. These can be used to allow users to explore the search results themselves.
- * @property searchEntryPoint Google search entry point for web searches. This contains an HTML/CSS
+ * @property searchEntryPoint Google Search entry point for web searches. This contains an HTML/CSS
  * snippet that **must** be embedded in an app to display a Google Search Entry point for follow-up
  * web searches related to the model's "Grounded Response".
  * @property groundingChunks The list of [GroundingChunk] classes. Each chunk represents a piece of